Change location_input from dropdown to lineEdit
From the last client meet we got to know that there
are around 400 possible location values, and it is
hard to use a dropdown input to select them based on
the current lab. The proposed update is to convert
that input into a lineEdit so we can just enter the
value.

Features:
* Add entry page have a lineEdit for the location input
* Check whether the location already exists under the lab
  and if it doesn't, insert the new location to the db
* Check the item is present in the location